groups treatment is not changed, though they do persist over fmol/showh and showq; planes now are grown as atoms an can be deleted in the same way, if created on grown structure, all symmetrical equivs are created; a change in groupable/selectable property, now only atoms and bonds can be grouped; covalent and Van Der Waals radii are added to the library and the VdW radii are used in calculation of voids etc; xyz format 'bug' is fixed (#177); planes generated by the same definition are now placed to the same graphics group for easier change of properties;

+++ gxlib/xbond.cpp
@@ -47,7 +47,7 @@ void TXBond::BondUpdated()  {
   else  {
     Params()[3] = C.Length();
     C.Normalise();
-    Params()[0] = (float)(acos(C[2])*180/M_PI);
+    Params()[0] = acos(C[2])*180/M_PI;
     if( olx_abs(Params()[0]-180) < 1e-3 )  { // degenerate case with Pi rotation
       Params()[1] = 0;
       Params()[2] = 1;
